% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_functions.R
\name{comp_expected_returns}
\alias{comp_expected_returns}
\title{Function to compute expected returns using a simple model selection approach}
\usage{
comp_expected_returns(returns, horizon)
}
\description{
Function to compute expected returns using a simple model selection approach
}
