% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{compute_sigma_0}
\alias{compute_sigma_0}
\title{Compute Sigma_0 p.93 Su and Wang (2017).}
\usage{
compute_sigma_0(res_set, iT, ip)
}
\arguments{
\item{res_set}{Residuals.}

\item{iT}{Number of time periods.}

\item{ip}{Number of assets.}
}
\value{
Sigma_0 from page 93 in Su and Wang (2017).
}
\description{
Compute Sigma_0 p.93 Su and Wang (2017).
}
\keyword{internal}
