% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{get_object_size}
\alias{get_object_size}
\title{the function will return the size of obj
and it is smart in the sense that it will choose the suitable unit}
\usage{
get_object_size(obj)
}
\arguments{
\item{obj}{Object}
}
\description{
the function will return the size of obj
and it is smart in the sense that it will choose the suitable unit
}
