% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{sqrt_matrix}
\alias{sqrt_matrix}
\title{Compute the Square Root of a Matrix}
\usage{
sqrt_matrix(Amat)
}
\arguments{
\item{Amat}{A numeric symmetric matrix.}
}
\value{
A matrix that is the square root of \code{Amat}.
}
\description{
Computes the square root of a symmetric matrix via eigen decomposition.
Negative eigenvalues are handled by taking the square root of their absolute values.
}
\keyword{internal}
