% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\docType{class}
\name{TVMVP}
\alias{TVMVP}
\title{Time Varying Minimum Variance Portfolio (TVMVP) Class}
\description{
This class implements a time-varying mean-variance portfolio model.
}
\section{Usage}{

```
```
}

\section{Arguments}{

* `data`: A tibble matrix for the data matrix.
}

\section{Methods}{

- `$new(data = NULL)`: Initialize the class.

- `$set_data(data)`: Set the data.

- `$get_data()`: Get the data.
}

