% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\name{TVMVP}
\alias{TVMVP}
\title{Time Varying Minimum Variance Portfolio (TVMVP) Class}
\description{
This class implements a time-varying mean-variance portfolio model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TVMVP-new}{\code{TVMVP$new()}}
\item \href{#method-TVMVP-set_data}{\code{TVMVP$set_data()}}
\item \href{#method-TVMVP-get_data}{\code{TVMVP$get_data()}}
\item \href{#method-TVMVP-clone}{\code{TVMVP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TVMVP-new"></a>}}
\if{latex}{\out{\hypertarget{method-TVMVP-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TVMVP$new(data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TVMVP-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-TVMVP-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TVMVP$set_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TVMVP-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-TVMVP-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TVMVP$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TVMVP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TVMVP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TVMVP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
